data-streaming/requirements.txt
# pandas
# requests
# pymongo
# jsonschema
# pika
# fitz
python-dotenv
